详情接口问题

+++ murder-mystery/src/main/java/cn/wyz/murdermystery/bean/MurderMystery.java
@@ -154,7 +154,7 @@ public class MurderMystery extends BaseEntity {
     /**
      * 查看次数
      */
-    private Long reviews;
+    private Integer reviews;
 
     private String location;
 }
